% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{deg2rad}
\alias{deg2rad}
\title{Pac-Man SVM}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Angle in degrees}
}
\value{
Pac-Man SVM
}
\description{
Conversion between degrees and radians
}
\keyword{machine-learning}
\keyword{visualization}
