% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsvm.r
\name{pac.lsvm}
\alias{pac.lsvm}
\title{Pac-Man SVM}
\usage{
pac.lsvm(
  x,
  y,
  l,
  title,
  axis_label,
  train_size = 0.7,
  rand_state = sample(1:2^15, 1)
)
}
\arguments{
\item{x, y}{Numeric data}

\item{l}{Numeric labels data}

\item{title}{Figure title}

\item{axis_label}{Label for the axis}

\item{train_size}{Fraction of total data that the SVM will train on}

\item{rand_state}{Value of the random state used to set the seed}
}
\value{
Pac-Man SVM
}
\description{
A visualization technique in R for regression analysis results, specifically residual values, based on a restricted
radial coordinate system. It provides a broad view perspective on the performance of regression models, and supports
most model inputs. See the pacviz documentation page for more information: https://pharaohcola13.github.io/pacviz/book/
}
\keyword{machine-learning}
\keyword{visualization}
