% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{pac.plot}
\alias{pac.plot}
\title{Pac-Man SVM}
\usage{
pac.plot(x, y, title, xaxislabel, yaxislabel, xunits, yunits, color1 = "gold")
}
\arguments{
\item{x, y}{Numeric data}

\item{title}{Figure title}

\item{xaxislabel}{Angular axis label}

\item{yaxislabel}{Radial axis label}

\item{xunits}{String to define units on the angular axis (For temperature measurements use 'degC' or 'degF')}

\item{yunits}{String to define units on the radial axis (For temperature measurements use 'degC' or 'degF')}

\item{color1}{Color value as string or rgb}
}
\value{
Pac-Man SVM
}
\description{
A visualization technique in R for regression analysis results, specifically residual values, based on a restricted
radial coordinate system. It provides a broad view perspective on the performance of regression models, and supports
most model inputs. See the pacviz documentation page for more information: https://pharaohcola13.github.io/pacviz/book/
}
\examples{
# Generic Pac-Man residual
data("cars")
pac.plot(cars$dist,cars$speed, 'Example 1', "Distance", "Speed", 'm', 'm/s')
}
\keyword{machine-learning}
\keyword{visualization}
