% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacman.r
\name{pacman}
\alias{pacman}
\title{Pac-Man Residual function}
\usage{
pacman(x, y, title)
}
\arguments{
\item{x:}{domain of the data y: range of the data title: title of the plot}
}
\description{
This function will create a pac-man residual plot for for regression analysis.
}
\examples{
pacman()
}
\keyword{regression}
