% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rad2deg}
\alias{rad2deg}
\title{Radian angle conversion}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Angle in radians}
}
\value{
Angle in degrees
}
\description{
Conversion between radians and degrees
}
\keyword{conversion}
