% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{unit_format}
\alias{unit_format}
\title{Unit formatting}
\usage{
unit_format(unit)
}
\arguments{
\item{unit}{Unit input}
}
\value{
A list of formatted units
}
\description{
Converts unit inputs into a format that can be displayed. Support is restricted to `degC`, `degF`.
}
