% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{deg2rad}
\alias{deg2rad}
\title{Degree angle conversion}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Angle in degrees}
}
\value{
Angle in radians
}
\description{
Conversion between degrees and radians
}
\keyword{conversion}
