% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{linMap}
\alias{linMap}
\title{Linear map}
\usage{
linMap(x, i, f)
}
\arguments{
\item{x}{Range of values to be mapped}

\item{i}{Lowest value}

\item{f}{Largest value}
}
\value{
A set of values spanning from i to f
}
\description{
A function that will map a range of values to a different set of values.
}
