% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{linMap}
\alias{linMap}
\title{Pac-Man SVM}
\usage{
linMap(x, i, f)
}
\arguments{
\item{x}{Range of values to be mapped}

\item{i}{Lowest value}

\item{f}{Largest value}
}
\value{
Pac-Man SVM
}
\description{
linear map
}
\keyword{machine-learning}
\keyword{visualization}
