% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsvm.r
\name{lsvm}
\alias{lsvm}
\title{Linear SVM with decision hyperplane and support vectors}
\usage{
lsvm(
  x,
  y,
  l,
  title,
  xaxis,
  yaxis,
  train_size = 0.7,
  rand_state = sample(1:2^15, 1)
)
}
\arguments{
\item{x, y}{Numeric data}

\item{l}{Numeric labels data}

\item{title}{Figure title}

\item{xaxis, yaxis}{Vector with the first entry being the axis label and the second entry being units}

\item{train_size}{Fraction of total data that the SVM will train on}

\item{rand_state}{Value of the random state used to set the seed}
}
\value{
Linear SVM plot
}
\description{
Generates a Linear Support Vector Machine and draws the decision hyperplane and support vectors
}
\keyword{machine-learning}
\keyword{visualization}
