% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.r
\name{pac.lsvm}
\alias{pac.lsvm}
\title{Pac-Man SVM}
\usage{
pac.lsvm(
  x,
  y,
  l,
  title,
  taxis,
  train_size = 0.7,
  rand_state = sample(1:2^15, 1)
)
}
\arguments{
\item{x, y}{Numeric data}

\item{l}{Numeric labels data}

\item{title}{Figure title}

\item{taxis}{Vector with the first entry being the axis label and the second entry being units}

\item{train_size}{Fraction of total data that the SVM will train on}

\item{rand_state}{Value of the random state used to set the seed}
}
\value{
Pac-Man SVM plot
}
\description{
A Pac-Man style SVM. (Under Development)
}
\keyword{machine-learning}
\keyword{visualization}
