% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{pac.plot}
\alias{pac.plot}
\title{Pac-Man plotting function}
\usage{
pac.plot(x, y, title, taxis, raxis, color1 = "gold")
}
\arguments{
\item{x, y}{Numeric data}

\item{title}{Figure title}

\item{taxis, raxis}{Vector with the first entry being the axis label and the second entry being units}

\item{color1}{Color value as string or rgb}
}
\value{
Pac-Man SVM
}
\description{
A method of plotting traditional Cartesian data, based on a restricted
radial coordinate system, while preserving the information.
}
\examples{
# Generic Pac-Man plot
data("cars")
pac.plot(cars$dist,cars$speed, 'Example 1', c("Distance", "m"), c("Speed", "m/s"))
}
\keyword{visualization}
