% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacman.r
\name{pacman}
\alias{pacman}
\title{Pac-Man Residual Function}
\usage{
pacman(
  x,
  y,
  title,
  unit,
  axis_label,
  model = lm(y ~ x, data = data.frame(x, y)),
  color1 = "Yellow",
  standardize = FALSE
)
}
\arguments{
\item{x, y}{Numeric data}

\item{title}{String}

\item{unit}{String to define units on the angular axis}

\item{axis_label}{String to label angular axis}

\item{model}{Ouput from lm()}

\item{color1}{Color strings}

\item{standardize}{Check to standardize the residual}
}
\value{
Pac-Man residual plot
}
\description{
This function will create a pac-man residual plot for for regression analysis. The data will run through a linear regression and plot the resulting factors of standard deviation the domain of the data.
}
\examples{
x <- rnorm(20, mean=0, sd=10)
y <- log(rnorm(20, mean=0, sd=10), base=exp(1))
pacman(x,y,'Title','units', 'Axis Label')
}
\keyword{regression}
\keyword{visualization}
