% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacman.r
\name{pacman}
\alias{pacman}
\title{Pac-Man Residual Function}
\usage{
pacman(
  x,
  y,
  title,
  unit,
  axis_label,
  model = lm(y ~ x, data = data.frame(x, y)),
  color1 = "gold",
  standardize = FALSE
)
}
\arguments{
\item{x, y}{Numeric data}

\item{title}{Figure title}

\item{unit}{String to define units on the angular axis (For temperature measurements use 'degC' or 'degF')}

\item{axis_label}{Angular axis label}

\item{model}{An object for which the extraction of model residuals is meaningful.}

\item{color1}{Color value as string or rgb}

\item{standardize}{Boolean to standardize the residual value}
}
\value{
Pac-Man residual plot
}
\description{
A visualization technique in R for regression analysis results, specifically residual values, based on a restricted
radial coordinate system. It provides a broad view perspective on the performance of regression models, and supports
most model inputs. See the pacviz documentation page for more information: https://pharaohcola13.github.io/pacviz/book/
}
\examples{
# Generic Pac-Man residual
x <- rnorm(20, mean=0, sd=10)
y <- log(rnorm(20, mean=0, sd=10), base=exp(1))
pacman(x,y,'Example 1','units', 'Axis Label')

# Pac-Man residual using alternate color, residual standardization, and temperature units
x <- rnorm(20, mean=0, sd=10)
y <- log(rnorm(20, mean=0, sd=10), base=exp(1))
pacman(x,y, 'Example 2', 'degC', "Temperature", color1="lightblue", standardize=TRUE)
}
\keyword{regression}
\keyword{visualization}
