% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacres.r
\name{pacres}
\alias{pacres}
\title{Pac-Man Residual Function}
\usage{
pacres(x, y, title, color1 = "Yellow", color2 = "White")
}
\arguments{
\item{x, y}{Numeric data}

\item{title}{String}

\item{color1, color2}{Color strings}
}
\value{
Pac-Man residual plot
}
\description{
This function will create a pac-man residual plot for for regression analysis. The data will run through a linear regression and plot the resulting factors of standard deviation against an arbitrary angular measurement.
}
\examples{
pacres(rnorm(20, mean=0, sd=1),exp(rnorm(20, mean=0, sd=1)),"Package_Test", "Yellow", "White")
}
\keyword{regression}
\keyword{visualization}
