% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacres.r
\name{pacres}
\alias{pacres}
\title{Pac-Man Residual function}
\usage{
pacres(x, y, title, color1, color2)
}
\arguments{
\item{x, y}{Numeric data}

\item{title}{String}

\item{color1, color2}{Color strings}
}
\value{
Pac-Man residual plot
}
\description{
This function will create a pac-man residual plot for for regression analysis. The data will run through a linear regression and plot the resulting factors of standard deviation against an arbitrary angular measurement.
}
\keyword{regression}
\keyword{visualization}
