% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rad2deg}
\alias{rad2deg}
\title{Pac-Man SVM}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Angle in radians}
}
\value{
Pac-Man SVM
}
\description{
Conversion between radians and degrees
}
\keyword{machine-learning}
\keyword{visualization}
