% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{svm.partition}
\alias{svm.partition}
\title{Pac-Man SVM}
\usage{
\method{svm}{partition}(x, y, l, train_size)
}
\arguments{
\item{x, y}{Numeric data}

\item{l}{Numeric labels data}

\item{train_size}{Fraction of total data that the SVM will train on}
}
\value{
Pac-Man SVM
}
\description{
A visualization technique in R for regression analysis results, specifically residual values, based on a restricted
radial coordinate system. It provides a broad view perspective on the performance of regression models, and supports
most model inputs. See the pacviz documentation page for more information: https://pharaohcola13.github.io/pacviz/book/
}
\keyword{machine-learning}
\keyword{visualization}
